Remove debugHandlers and use hjandlers directly in debug page
